ph and tds input with ads1115

+++ esp8266/lib/MQTTModule/MQTTModule.cpp
@@ -91,10 +91,10 @@ void MQTTModule::begin()
     connectToWifi();
 }
 
-void MQTTModule::publishData(const char *topic, int tdsValue, float phValue, float tempValue)
+void MQTTModule::publishData(const char *topic, float tdsValue, float phValue, float tempValue)
 {
     char influxDBLineProtocol[128]; // Adjust buffer size as per your requirements
-    sprintf(influxDBLineProtocol, "esp0,hydro_number=0 tds_value=%d,ph_value=%f,temp_value=%f", tdsValue, phValue, tempValue);
+    sprintf(influxDBLineProtocol, "esp0,hydro_number=0 tds_value=%f,ph_value=%f,temp_value=%f", tdsValue, phValue, tempValue);
     publish(topic, influxDBLineProtocol);
 }
